+++ client/components/UI/Post/MarketPost.tsx
@@ -15,19 +15,18 @@ function MarketPost({ imgSource, title, className, path, price }: PostProps) {
   return (
     <div
       className={twMerge(
-        ' bg-lightPink flex flex-row py-2 px-4 mb-2 mx-5 rounded-sm drop-shadow-xl',
+        ' bg-lightPink py-2 mb-2 rounded-sm drop-shadow-xl',
         className
       )}
     >
-      <img className="h-14 w-14" src={imgSource} alt={title} />
+      <img className="w-full h-28" src={imgSource} alt={title} />
       <div className="flex flex-col pl-4 pt-2">
-        <p className="text-black font-bold font-xl">
+        <div className="font-bold text-grey font-base">
+          <p>${price}</p>
+        </div>
+        <p className="font-xl font-light">
           <Link to={path}>{title}</Link>
         </p>
-        <div className="flex flex-row space-x-10 text-grey font-base">
-          <p>${price}</p>
-          {/* <p>({comment}) comments</p> */}
-        </div>
       </div>
     </div>
   )
